fix: stars with TanStack query
The 'starred' attribute is now computed only when pulls are fetched, now that we are using TanStack query. It means that (un)starring a pull request was basically broken since the move to TanStack query. This change remove the 'starred' attribute from the Pull type, and determines on-the-fly whether a pull request is starred.